requirement.txt
tensorflow==2.18.0
numpy==1.24.3
opencv-python==4.8.0.74
Pillow==10.0.1
tkintertable==1.3.2
